% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeting.R
\name{greet}
\alias{greet}
\title{Hello world from CGO via Rcpp}
\usage{
greet(message = "World")
}
\arguments{
\item{message}{Character vector}
}
\value{
Character vector
}
\description{
An example function.
}
