% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCgoHello.R
\docType{package}
\name{RCgoHello-package}
\alias{RCgoHello}
\alias{RCgoHello-package}
\title{RCgoHello: Tiny Example Using CGO via 'Rcpp'}
\description{
A small example that shows how to use CGO-built shared libraries
  from Rcpp function.
}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
\keyword{internal}
