% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCgo-package.R
\docType{package}
\name{RcppCgo-package}
\alias{RcppCgo}
\alias{RcppCgo-package}
\title{RcppCgo: Tiny Example Using CGO via 'Rcpp'}
\description{
A small example that shows how to use CGO-built archive from Rcpp.
}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
\keyword{internal}
